requirements.txt
RPi.GPIO==0.7.1
picamera2
numpy
PyYAML
pytest
